+++ CalcGtinExamples/MainWindow.xaml.cs
@@ -1,13 +1,16 @@
 ﻿using System.Windows;
+using CalcGtinExamplesCore;
 
 namespace CalcGtinExamples
 {
     /// <summary>
     /// Interaction logic for MainWindow.xaml
     /// </summary>
-    public partial class MainWindow : Window
+    // ReSharper disable once UnusedMember.Global
+    public partial class MainWindow
     {
         private readonly ViewModel _viewModel;
+        private readonly BarcodeHelper _barcode = BarcodeHelper.Instance;
 
         public MainWindow()
         {
@@ -23,22 +26,22 @@ namespace CalcGtinExamples
         private void ButtonEanClear_OnClick(object sender, RoutedEventArgs e)
         {
             _viewModel.ClearEan();
-            _viewModel.Gtin = _viewModel.GetGtin(_viewModel.Ean, EnumGtinVariant.Var1);
+            _viewModel.Gtin = _barcode.GetGtin(_viewModel.Ean, EnumGtinVariant.Var1);
         }
 
         private void ButtonGtinCalcV1_OnClick(object sender, RoutedEventArgs e)
         {
-            _viewModel.Gtin = _viewModel.GetGtin(_viewModel.Ean, EnumGtinVariant.Var1);
+            _viewModel.Gtin = _barcode.GetGtin(_viewModel.Ean, EnumGtinVariant.Var1);
         }
         
         private void ButtonGtinCalcV2_OnClick(object sender, RoutedEventArgs e)
         {
-            _viewModel.Gtin = _viewModel.GetGtin(_viewModel.Ean, EnumGtinVariant.Var2);
+            _viewModel.Gtin = _barcode.GetGtin(_viewModel.Ean, EnumGtinVariant.Var2);
         }
         
         private void ButtonGtinCalcV3_OnClick(object sender, RoutedEventArgs e)
         {
-            _viewModel.Gtin = _viewModel.GetGtin(_viewModel.Ean, EnumGtinVariant.Var3);
+            _viewModel.Gtin = _barcode.GetGtin(_viewModel.Ean);
         }
     }
 }